Fully Document the Organizations Frontend Feature
This commit adds documentation to the entire organization front end feature, including components, services, and more.
