+++ frontend/src/app/admin/organization/list/admin-organization-list.component.ts
@@ -10,7 +10,7 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
 import { permissionGuard } from 'src/app/permission.guard';
-import { Organization } from '../../../organization/organization.service';
+import { Organization } from '../../../organization/organization.model';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { AdminOrganizationService } from '../admin-organization.service';
 import { Observable } from 'rxjs';
