+++ frontend/src/app/organization/organization-filter/organization-filter.pipe.ts
@@ -7,7 +7,7 @@
  */
 
 import { Pipe, PipeTransform } from '@angular/core';
-import { Organization } from '../organization.service';
+import { Organization } from '../organization.model';
 
 @Pipe({
   name: 'organizationFilter'
