+++ frontend/src/app/organization/organization-routing.module.ts
@@ -1,3 +1,12 @@
+/**
+ * The Organization Routing Module holds all of the routes that are children
+ * to the path /organizations/...
+ * 
+ * @author Ajay Gandecha, Jade Keegan, Brianna Ta, Audrey Toney
+ * @copyright 2023
+ * @license MIT
+ */
+
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { OrganizationPageComponent } from './organization-page/organization-page.component';
