+++ frontend/src/app/organization/organization.resolver.ts
@@ -9,12 +9,15 @@
 
 import { inject } from "@angular/core";
 import { ResolveFn } from "@angular/router";
-import { Organization, OrganizationService } from "./organization.service"
+import { Organization } from "./organization.model";
+import { OrganizationService } from "./organization.service"
 
+/** This resolver injects the list of organizations into the organization component. */
 export const organizationResolver: ResolveFn<Organization[] | undefined> = (route, state) => {
     return inject(OrganizationService).getOrganizations();
 };
 
+/** This resolver injects an organization into the organization detail component. */
 export const organizationDetailResolver: ResolveFn<Organization | undefined> = (route, state) => {
     if(route.paramMap.get('slug')! != "new") {
         return inject(OrganizationService).getOrganization(route.paramMap.get('slug')!);
