+++ frontend/src/app/organization/widgets/organization-details-info-card/organization-details-info-card.widget.ts
@@ -1,8 +1,17 @@
+/**
+ * The Organization Details Info Card widget abstracts the implementation of each
+ * individual organization detail card from the whole organization detail page.
+ * 
+ * @author Ajay Gandecha, Jade Keegan, Brianna Ta, Audrey Toney
+ * @copyright 2023
+ * @license MIT
+ */
+
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { Organization } from '../../organization.service';
+import { Organization } from '../../organization.model';
 import { Profile } from 'src/app/profile/profile.service';
 import { PermissionService } from 'src/app/permission.service';
 import { Observable } from 'rxjs';
@@ -14,32 +23,40 @@ import { Observable } from 'rxjs';
 })
 export class OrganizationDetailsInfoCard implements OnInit, OnDestroy {
 
+    /** The organization to show */
     @Input() organization?: Organization;
+    /** The currently logged in user */
     @Input() profile?: Profile;
 
+    /** Holds data on whether or not the user is on a mobile device */
     public isHandset: boolean = false;
     private isHandsetSubscription!: Subscription;
 
+    /** Holds data on whether or not the user is on a tablet */
     public isTablet: boolean = false;
     private isTabletSubscription!: Subscription;
 
     /** Stores whether the user has admin permission over the current organization. */
     public adminPermission$: Observable<boolean>;
 
+    /** Constructs the organization detail info card widget */
     constructor(private breakpointObserver: BreakpointObserver, private permission: PermissionService) {
         this.adminPermission$ = this.permission.check('admin.view', 'admin/');
     }
 
+    /** Runs whenever the view is rendered initally on the screen */
     ngOnInit(): void {
         this.isHandsetSubscription = this.initHandset();
         this.isTabletSubscription = this.initTablet();
     }
 
+    /** Unsubscribe from subscribers when the page is destroyed */
     ngOnDestroy(): void {
         this.isHandsetSubscription.unsubscribe();
         this.isTabletSubscription.unsubscribe();
     }
 
+    /** Determines whether the page is being used on a mobile device */
     private initHandset() {
         return this.breakpointObserver
             .observe([Breakpoints.Handset, Breakpoints.TabletPortrait])
@@ -47,6 +64,7 @@ export class OrganizationDetailsInfoCard implements OnInit, OnDestroy {
             .subscribe(isHandset => this.isHandset = isHandset);
     }
 
+    /** Determines whether the page is being used on a tablet */
     private initTablet() {
         return this.breakpointObserver
             .observe(Breakpoints.TabletLandscape)
